feat: 클라코 취향 컴포넌트 생성
